Add `write_through` to ActiveSupport cache adapter
The current approach with caching adapters does something like this:

1. On write, delete the entry from the cache.
2. On the next read, repopulate the cache.

Unfortunately this can be subject to subtle race conditions if the
underlying storage (in our case, a replicated database) gives a stale
read for item 2 - which may be in a different process / thread / host
entirely. More details on our specific case are available at
https://gitlab.com/gitlab-org/gitlab/-/issues/325452.

The new `write_through` option to the ActiveSupportCacheStore adapter's
initialiser allows the user to opt in to write-through caching, where
the new value is immediately written to the cache in step 1.